# examples/cursor/web/openapi.json
{"openapi":"3.1.0","info":{"title":"Cursor MCP Server","description":"MCP Server for Cursor integration with MemMachine","version":"1.0.0"},"paths":{"/api/health":{"get":{"summary":"Health Check","description":"Simple health check endpoint.","operationId":"health_check_api_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseSchema"}}}}}}},"/api/debug":{"get":{"summary":"Debug Info","description":"Debug endpoint to show server configuration.","operationId":"debug_info_api_debug_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseSchema"}}}}}}},"/api/auth/register":{"post":{"summary":"Register","description":"Register a new user.","operationId":"register_api_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegistration"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseSchema"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/login":{"post":{"summary":"Login","description":"Login with username/password to get a bearer token.","operationId":"login_api_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLogin"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseSchema"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/logout":{"post":{"summary":"Logout","description":"Logout by revoking the provided token.","operationId":"logout_api_auth_logout_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseSchema"}}}}}}},"/api/auth/me":{"get":{"summary":"Get Current User Info","description":"Get current authenticated user information.","operationId":"get_current_user_info_api_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseSchema"}}}}}}},"/api/memory/get_profile_memory":{"get":{"summary":"Rest Get Profile Memory","description":"Get the profile memory for the current session via REST endpoint.\n\nThis endpoint reuses the shared helper to fetch profile memory using the\nresolved session id (explicit parameter takes precedence over headers).","operationId":"rest_get_profile_memory_api_memory_get_profile_memory_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":10,"title":"Limit"}},{"name":"session_id","in":"query","required":false,"schema":{"type":"string","default":"","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseSchema"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/memory/get_episodic_memory":{"get":{"summary":"Rest Get Episodic Memory","description":"Get the episodic memory for the current session via REST endpoint.\n\nThis endpoint reuses the shared helper to fetch episodic memory using the\nresolved session id (explicit parameter takes precedence over headers).","operationId":"rest_get_episodic_memory_api_memory_get_episodic_memory_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":10,"title":"Limit"}},{"name":"session_id","in":"query","required":false,"schema":{"type":"string","default":"","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseSchema"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/memory/add":{"post":{"summary":"Rest Add Memory","description":"REST endpoint to add a memory episode (mirrors MCP add_memory).","operationId":"rest_add_memory_api_memory_add_post","parameters":[{"name":"content","in":"query","required":true,"schema":{"type":"string","title":"Content"}},{"name":"session_id","in":"query","required":true,"schema":{"type":"string","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseSchema"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/memory/search":{"get":{"summary":"Rest Search Memory","description":"REST endpoint to search memories (mirrors MCP search_memory).","operationId":"rest_search_memory_api_memory_search_get","parameters":[{"name":"query","in":"query","required":true,"schema":{"type":"string","title":"Query"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":5,"title":"Limit"}},{"name":"session_id","in":"query","required":false,"schema":{"type":"string","default":"","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseSchema"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/memory/episodic/delete":{"delete":{"summary":"Rest Delete Session Memory","description":"REST endpoint to delete all memories for the current session (mirrors MCP delete_session_memory).","operationId":"rest_delete_session_memory_api_memory_episodic_delete_delete","parameters":[{"name":"session_id","in":"query","required":true,"schema":{"type":"string","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseSchema"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/sessions":{"get":{"summary":"Get Current User Sessions","description":"Get all sessions for the current authenticated user.\n\nThis endpoint fetches all sessions associated with the current\nauthenticated user from the MemMachine backend.","operationId":"get_current_user_sessions_api_sessions_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseSchema"}}}}}}},"/api/mcp/tools":{"get":{"summary":"List Mcp Tools","description":"List all registered MCP tools.\n\nThis implementation inspects module globals for functions with the\n\"mcp_\" prefix, which mirrors the set of MCP-exposed tools in this\nserver. It avoids relying on FastMCP private internals.","operationId":"list_mcp_tools_api_mcp_tools_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response List Mcp Tools Api Mcp Tools Get"}}}}}}},"/{full_path}":{"get":{"summary":"Serve Spa","description":"Serve SPA for all non-API routes.","operationId":"serve_spa__full_path__get","parameters":[{"name":"full_path","in":"path","required":true,"schema":{"type":"string","title":"Full Path"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ResponseSchema":{"properties":{"success":{"type":"boolean","title":"Success","description":"Request success status","default":true},"message":{"type":"string","title":"Message","description":"Response message","default":"Success"},"data":{"anyOf":[{},{"type":"null"}],"title":"Data","description":"Response data"},"error_code":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Error Code","description":"Error code if failed"}},"type":"object","title":"ResponseSchema","description":"Standard API response schema"},"UserLogin":{"properties":{"username":{"type":"string","title":"Username","description":"Username","example":"john_doe"},"password":{"type":"string","title":"Password","description":"Password","example":"mySecurePass123"}},"type":"object","required":["username","password"],"title":"UserLogin","description":"User login request model."},"UserRegistration":{"properties":{"username":{"type":"string","maxLength":50,"minLength":3,"title":"Username","description":"Username","example":"john_doe"},"email":{"anyOf":[{"type":"string","format":"email"},{"type":"null"}],"title":"Email","description":"Email address","example":"john@example.com"},"password":{"type":"string","minLength":6,"title":"Password","description":"Password","example":"mySecurePass123"}},"type":"object","required":["username","password"],"title":"UserRegistration","description":"User registration request model."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"Bearer":{"type":"http","scheme":"bearer","bearerFormat":"JWT","description":"Enter your authentication token"}}},"security":[{"Bearer":[]}]}